feat(courses): add global exception handler

+++ src/main/java/com/weleson/courses_api/courses/useCases/DeteleCourseUseCase.java
@@ -6,6 +6,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import com.weleson.courses_api.courses.repositories.CourseRepository;
+import com.weleson.courses_api.exeptions.CourseNotFoundExeption;
 
 @Service
 public class DeteleCourseUseCase {
@@ -15,6 +16,9 @@ public class DeteleCourseUseCase {
 
   public void execute(UUID id) {
     try {
+      if (courseRepository.findById(id).isEmpty()) {
+        throw new CourseNotFoundExeption();
+      }
       courseRepository.deleteById(id);
     } catch (Exception e) {
       throw new RuntimeException("Internal server error");
